refactor: trashs -> trash로 변경
